chore: infra

+++ synctv-core/src/bootstrap/user.rs
@@ -38,7 +38,7 @@ pub async fn bootstrap_root_user(pool: &PgPool, config: &BootstrapConfig, develo
     let root_exists = sqlx::query_scalar::<_, bool>(
         "SELECT EXISTS(SELECT 1 FROM users WHERE role = $1 AND deleted_at IS NULL LIMIT 1)"
     )
-    .bind(UserRole::Root as i16)
+    .bind(UserRole::Root)
     .fetch_one(pool)
     .await?;
